src/tasks/vision/05-image-to-text/requirements.txt
transformers
torch
Pillow